+++ app/controllers/spam2_controller.rb
@@ -4,7 +4,6 @@ class Spam2Controller < ApplicationController
   def _spam
     if logged_in_as(%w(moderator admin))
       @nodes = Node.paginate(page: params[:page], per_page: params[:pagination])
-      @queue_tag = TagSelection.where(following: true, user_id: current_user.id).limit(10)
       @nodes = case params[:type]
                  when 'wiki'
                    @nodes.where(type: 'page', status: 1).order('changed DESC')
@@ -14,9 +13,6 @@ class Spam2Controller < ApplicationController
                    @nodes.where(status: 0).order('changed DESC')
                  when 'created'
                    @nodes.where(status: [0, 4]).order('created DESC')
-                 when 'queue'
-                   current_user.moderator_queue.order('created DESC')
-                                 .paginate(page: params[:page], per_page: params[:pagination])
                  else
                    @nodes.where(status: [0, 4]).order('changed DESC')
                  end
@@ -42,6 +38,33 @@ class Spam2Controller < ApplicationController
                  @flags.where(type: 'node')
                else
                  @flags
+               end
+      render template: 'spam2/_spam'
+    else
+      flash[:error] = 'Only moderators can moderate posts.'
+      redirect_to '/dashboard'
+    end
+  end
+
+  def _spam_queue
+    if logged_in_as(%w(moderator admin))
+      @tag_queue = TagSelection.where(following: true, user_id: current_user.id)
+      nodes = []
+      @tag_queue.each do |tag_queue_name|
+        nodes += NodeTag.where(tid: tag_queue_name.tid).collect(&:nid)
+      end
+      @queue = Node.where(nid: nodes, type: %w(note page))
+                   .paginate(page: params[:page], per_page: params[:pagination])
+                   .order('changed DESC').distinct
+      @queue = case params[:type]
+               when 'unmoderated'
+                 @queue.where(status: 4).order('created DESC')
+               when 'spammed'
+                 @queue.where(status: 0).order('created DESC')
+               when 'flagged'
+                 @queue.where('flag > ?', 0).order('flag DESC')
+               else
+                 @queue.where(status: [0, 4]).order('changed DESC')
                end
       render template: 'spam2/_spam'
     else
@@ -104,126 +127,6 @@ class Spam2Controller < ApplicationController
     end
   end
 
-  def batch_spam
-    if logged_in_as(%w(moderator admin))
-      user_spamed = []
-      node_spamed = 0
-      params[:ids].split(',').uniq.each do |nid|
-        node = Node.find nid
-        node_spamed += 1
-        node.spam
-        user = node.author
-        user_spamed << user.id
-        user.ban
-      end
-      flash[:notice] = node_spamed.to_s + ' nodes spammed and ' + user_spamed.length.to_s + ' users banned.'
-      redirect_to '/spam2'
-    else
-      flash[:error] = 'Only admins and moderators can mark a batch spam.'
-      redirect_to '/dashboard'
-    end
-  end
-
-  def batch_publish
-    if logged_in_as(%w(moderator admin))
-      node_published = 0
-      user_published = []
-      params[:ids].split(',').uniq.each do |nid|
-        node = Node.find nid
-        node_published += 1
-        node.publish
-        user = node.author
-        user.unban
-        user_published << user.id
-      end
-      flash[:notice] = node_published.to_s + ' nodes published and ' + user_published.length.to_s + ' users unbanned.'
-      redirect_to '/spam2'
-    else
-      flash[:error] = 'Only admins and moderators can batch publish.'
-      redirect_to '/dashboard'
-    end
-  end
-
-  def batch_delete
-    if logged_in_as(%w(moderator admin))
-      node_delete = 0
-      params[:ids].split(',').uniq.each do |nid|
-        node = Node.find nid
-        node_delete += 1
-        node.delete
-      end
-      flash[:notice] = node_delete.to_s + ' nodes deleted'
-      redirect_back fallback_location: root_path
-    else
-      flash[:error] = 'Only admins and moderators can batch delete.'
-      redirect_to '/dashboard'
-    end
-  end
-
-  def batch_ban
-    if logged_in_as(%w(admin moderator))
-      user_ban = []
-      params[:ids].split(',').uniq.each do |nid|
-        node = Node.find nid
-        user = node.author
-        user_ban << user.id
-        user.ban
-      end
-      flash[:notice] = user_ban.length.to_s + ' users banned.'
-      redirect_back fallback_location: root_path
-    else
-      flash[:error] = 'Only admins and moderators can ban users.'
-      redirect_to '/dashboard'
-    end
-  end
-
-  def batch_unban
-    unbanned_users = []
-    if logged_in_as(%w(moderator admin))
-      params[:ids].split(',').uniq.each do |node_id|
-        node = Node.find node_id
-        user_unban = node.author
-        unbanned_users << user_unban.id
-        user_unban.unban
-      end
-      flash[:notice] = unbanned_users.length.to_s + ' users unbanned.'
-      redirect_back fallback_location: root_path
-    else
-      flash[:error] = 'Only admins and moderators can unban users.'
-      redirect_to '/dashboard'
-    end
-  end
-
-  def batch_ban_user
-    users = []
-    if logged_in_as(%w(admin moderator))
-      params[:ids].split(',').uniq.each do |user_id|
-        user_ban = User.find user_id
-        users << user_ban.id
-        user_ban.ban
-      end
-      flash[:notice] = users.length.to_s + ' users banned.'
-      redirect_back fallback_location: root_path
-    else
-      flash[:error] = 'Only moderators can moderate users.'
-      redirect_to '/dashboard'
-    end
-  end
-
-  def batch_unban_user
-    if logged_in_as(%w(moderator admin))
-      params[:ids].split(',').uniq.each do |id|
-        unban_user = User.find id
-        unban_user.unban
-      end
-      flash[:notice] = 'Success! users unbanned.'
-      redirect_back fallback_location: root_path
-    else
-      flash[:error] = 'Only admins and moderators can moderate users.'
-      redirect_to '/dashboard'
-    end
-  end
-
   def flag_node
     @node = Node.find params[:id]
     @node.flag_node
@@ -258,36 +161,6 @@ class Spam2Controller < ApplicationController
       @comment.unflag_comment
     else
       flash[:error] = 'Only moderators can unflag comments.'
-      redirect_to '/dashboard'
-    end
-  end
-
-  def batch_comment
-    if logged_in_as(%w(moderator admin))
-      comment_total = 0
-      user_total = []
-      params[:ids].split(',').uniq.each do |cid|
-        comment = Comment.find cid
-        comment_total += 1
-        user = comment.author
-        user_total << user.id
-        case params[:type]
-        when 'publish'
-          comment.publish
-          user.unban
-        when 'spam'
-          comment.spam
-          user.ban
-        when 'delete'
-          comment.delete
-        else
-          flash[:notice] = 'Invalid Url'
-        end
-      end
-      flash[:notice] = comment_total.to_s + ' comment and ' + user_total.length.to_s + ' user moderated'
-      redirect_back fallback_location: root_path
-    else
-      flash[:error] = 'Only admins and moderators can moderate comments and users.'
       redirect_to '/dashboard'
     end
   end
