+++ app/views/spam2/_spam.html.erb
@@ -7,40 +7,40 @@
 <div style="display:none;" id="nspam" class="alert alert-danger alert-dismissible spam-alert" role="alert">
   Content spammed
 </div>
-<div style="display:none;" id="nunflag" class="alert alert-danger alert-dismissible spam-alert" role="alert">
+<div style="display:none;" id="nunflag" class="alert alert-success alert-dismissible spam-alert" role="alert">
   Content unflagged
 </div>
 
 <div class="nav_top flex d-inline">
-  <p class="text-secondary pt-2 float-left nav_title h4">
+  <p class="text-secondary pt-2 float-left nav_title ">
     <% if params[:type] == "wiki" %>
-      <a href="/spam2" class="text-dark"><i class=" pt-1 pr-2 fa fa-certificate text-success"></i> Spam2 / </a><b> Wiki
-    <% elsif params[:type] == "queue"%>
-      <a href="/spam2" class="text-dark"><i class=" pt-1 pr-2 fa fa-clock text-success"></i> Spam2 / </a><b> Queue   
+      <a href="/spam2" class="text-dark h4"><i class=" pt-1 pr-2 fa fa-certificate text-success"></i> Spam2 / </a><b> <span class="h2">Wiki</span>
+    <% elsif params[:action] == "_spam_queue"%>
+      <a href="/spam2" class="text-dark h4"><i class=" pt-1 pr-2 fa fa-clock text-success"></i> Spam2 / </a><b> <span class="h2">Queue</span>
     <% elsif params[:action] == "_spam_revisions"%>
-      <a href="/spam2" class="text-dark"><i class=" pt-1 pr-2 fa fa-file text-primary"></i> Spam2 / </a><b> Revisions
+      <a href="/spam2" class="text-dark h4"><i class=" pt-1 pr-2 fa fa-file text-primary"></i> Spam2 / </a><b> <span class="h2">Revisions</span>
     <% elsif params[:action] == "_spam_comments"%>
-      <a href="/spam2" class="text-dark"><i class=" pt-1 pr-2 fa fa-comments text-info"></i> Spam2 / </a><b> Comments
+      <a href="/spam2" class="text-dark h4"><i class=" pt-1 pr-2 fa fa-comments text-info"></i> Spam2 / </a><b> <span class="h3">Comments</span>
     <% elsif params[:action] == "_spam_flags"%>
-      <a href="/spam2" class="text-dark"><i class=" pt-1 pr-2 fa fa-flag text-warning"></i> Spam2 / </a><b> Flags
+      <a href="/spam2" class="text-dark h4"><i class=" pt-1 pr-2 fa fa-flag text-warning"></i> Spam2 / </a><b> <span class="h2">Flags</span>
     <% elsif params[:action] == "_spam_users"%>
-      <a href="/spam2" class="text-dark"><i class=" pt-1 pr-2 fa fa-user text-info"></i> Spam2 / </a><b> Users
+      <a href="/spam2" class="text-dark h4"><i class=" pt-1 pr-2 fa fa-user text-info"></i> Spam2 / </a><b> <span class="h2">Users</span>
     <% else %>
-      <a href="/spam2" class="text-dark"><i class=" pt-1 pr-2 fa fa-check-circle text-success"></i> Spam2 / </a><b> Bulk Moderation 
+      <a href="/spam2" class="text-dark h4"><i class=" pt-1 pr-2 fa fa-check-circle text-success"></i> Spam2 / </a><b> <span class="h3">Moderation</span> 
     <% end %> </b>
   </p>
   <div class="dropdown float-right">
-    <a class="btn btn-light border border-curve" type="button" id="dropdownMenuLink" href="#" >
+    <a class="btn btn-light border border-curve mt-2" type="button" id="dropdownMenuLink" href="#" >
       Menu <i class="fa fa-align-left text-secondary" aria-hidden="true"></i>
     </a>
     <div class="dropdown-menu dropdown-menu-right">
-      <h6 class="dropdown-header">Menu</h6>
+      <h6 class="dropdown-header">Menu, Role: <%= current_user.role%> </h6>
         <a class="dropdown-item" href="/dashboard"><i class="fa fa-list-alt text-primary"></i> Dashboard</a>
       <div class="dropdown-divider"></div>  
         <h6 class="dropdown-header">Spam Moderation</h6>
-        <a class="dropdown-item <% unless  params[:action] =="_spam_comments" ||params[:type] =="queue" || params[:action] =="_spam_revisions" || params[:action] =="_spam_flags" || params[:type] == "wiki" || params[:action] == "_spam_users"%> active <% end %>" href="/spam2/filter/all/30"><i class="drop fa fa-address-book text-primary"></i> Bulk Moderation</a>
+        <a class="dropdown-item <% unless  params[:action] =="_spam_comments" ||params[:action] =="_spam_queue" || params[:action] =="_spam_revisions" || params[:action] =="_spam_flags" || params[:type] == "wiki" || params[:action] == "_spam_users"%> active <% end %>" href="/spam2/filter/all/30"><i class="drop fa fa-address-book text-primary"></i> Moderation</a>
         <a class="dropdown-item <% if params[:type] == "wiki" %> active <% end %>" id="wiki_drop" href="/spam2/filter/wiki/30" ><i class="drop fa fa-book text-primary"></i> Wiki</a>
-        <a class="dropdown-item <% if params[:type] == "queue" %> active <% end %>" id="queue_drop" href="/spam2/filter/queue/30" ><i class="drop fa fa-clock text-primary"></i> Queue</a>
+        <a class="dropdown-item <% if params[:action] == "_spam_queue" %> active <% end %>" href="/spam2/queue/filter/all/30" ><i class="drop fa fa-clock text-primary"></i> Queue</a>
         <a class="dropdown-item <% if params[:action] == "_spam_revisions" %> active <% end %>" href="/spam2/revisions"><i class="drop fa fa-list text-primary"></i> Revision</a>
         <a class="dropdown-item <% if params[:action] == "_spam_comments" %> active <% end %>" href="/spam2/comments/filter/all/30" ><i class="drop fa fa-comment text-primary"></i> Comments</a>
         <a class="dropdown-item <% if params[:action] == "_spam_flags" %> active <% end %>"  href="/spam2/flags/filter/all/30" ><i class="drop fa fa-flag text-primary"></i> Flag </a>
@@ -52,10 +52,6 @@
   $('#wiki_drop').on('click', function () { // Wiki filter
     localStorage.setItem('page-select', "30");
     localStorage.setItem('filter', 'wiki');
-  });
-  $('#queue_drop').on('click', function () { // Queue filter
-    localStorage.setItem('page-select', "30");
-    localStorage.setItem('filter', 'queue');
   });
 </script>
 <div class="d-flex w-100 " style="overflow-x:auto !important;">
@@ -120,18 +116,9 @@
 </div>
 <% end %>
 
-<% if params[:type] == "queue" %>
-<div class="py-2 pl-3 w-100">
-  <i class="fa fa-certificate text-success"></i>
-  <% if @queue_tag.length == 10 %> <span class="badge text-info"> + </span><% end %>
-  <% @queue_tag.each do |tag| %>
-    <span class="badge badge-pill badge-info"><%= tag.tagname %> <%= Tag.followers(tag.tagname).where('rusers.role = ?', 'moderator').or(Tag.followers(tag.tagname).where('rusers.role = ?', 'admin')).length %></span>  
-  <% end %>
-  </div>
-<% end %>
-
 <%= render partial: 'spam2/nodes' if @nodes %>
 <%= render partial: 'spam2/revisions' if @revisions %>
 <%= render partial: 'spam2/comments' if @comments %>
+<%= render partial: 'spam2/queue' if @queue %>
 <%= render partial: 'spam2/flags' if @flags %>
 <%= render partial: 'spam2/users' if @users %>
